+++ server.js
@@ -1,5 +1,5 @@
 const server = require('./index');
-const PORT = process.argv[2] || 3000;
+const PORT = process.argv[2] || 8080;
 
 server.listen(PORT, function() {
     console.log('listening on *:'+PORT)
